Close SQLite OPFS handles on worker shutdown
