+++ .changeset/green-birds-close.md
@@ -1,0 +1,5 @@
+---
+"@effect/sql-sqlite-wasm": patch
+---
+
+Close OPFS access handles when the SQLite worker shuts down.
